+++ packages/ds/src/shell/micro/CommandPaletteShell/index.ts
@@ -1,0 +1,6 @@
+/**
+ * CommandPaletteShell - ⌘K Command Interface
+ */
+
+export { CommandPaletteShell } from './CommandPaletteShell';
+export type { CommandPaletteShellProps, Command } from './CommandPaletteShell';